requirements.txt
flask
flask_mail
itsdangerous
flask_wtf
wtforms
flask_socketio
pymysql
email_validator
openai
ajax
simple-websocket
google-auth==2.3.2
google-auth-oauthlib
MarkupSafe
oauthlib
requests
requests-oauthlib
cachetools
certifi
chardet
urllib3
